feat(referrals): update referral components and replace ReferralDashboard with ReferralHubView (#517)
- Added support for "swap" transaction type in getTransactionHistoryTypeLabel function.
- Replaced ReferralDashboard with ReferralHubView in MobileDropdown and WalletDetails components for improved referral management.
- Removed unused ReferralDashboard and ReferralDashboardViewSkeleton components to streamline the codebase.
- Enhanced ReferralCTA to directly trigger the new referral view, improving user navigation.
